Добавил пин управления
