+++ js/data.js
@@ -1,4 +1,4 @@
-import { cards } from './map.js';
+import { createPopupCards } from './create-popup.js';
 
 const SERVER_PATH = 'https://23.javascript.pages.academy/keksobooking/data';
 
@@ -6,7 +6,7 @@ fetch(SERVER_PATH)
   .then((response) => response.json())
   .then((data) => {
 
-    cards(data);
+    createPopupCards(data);
 
   });
