+++ js/main.js
@@ -1,3 +1,3 @@
 import './validation.js';
 import './map.js';
-import './data.js';//Validation
+import './data.js';
